requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
redis==5.2.1
requests==2.32.3
Flask-Cors==5.0.0
python-dotenv==1.0.1